// src/app/models/IPlanProd.ts
import { IControlLote, ILinea } from ".";

export interface IPlanProd {
  cantidad?: number | null;
  cantidadProducida?: string | null;
  cantidadRechazos?: number | null;
  capacidad?: string | null;
  codigoModelo?: string | null;
  codigoModeloPar?: string | null;
  desde?: number | null;
  dobladora?: string | null;
  estado?: string | null;
  fecha?: Date | null;
  fechaInicio?: Date | null;
  fechaFinal?: Date | null;
  fpy?: string | null;
  hasta?: number | null;
  idLinea?: number | null;
  idModelo?: number | null;
  idModeloNavigation?: any | null;
  idPauta?: number | null;
  idProduccion?: number | null;
  idProveedor?: number | null;
  inicio?: any | null;
  insercion?: string | null;
  linea?: ILinea | null;
  lote?: string | null;
  loteCerrado?: string | null;
  mes?: number | null;
  nombre?: string | null;
  numeroOp?: string | null;
  opsMainDisplay?: any[];
  organizacion?: string | null;
  pendiente?: number | null;
  producido?: string | null;
  rechazados?: number | null;
  rechazos?: IControlLote[] | [];
  rechazosDeLote?: number;
  rutaLogo?: string | null;
  semielaborado?: any;
  semiElaboradoId?: number | null;
  target?: number | null;
  tipoSemiElaborado?: string;
  tipoUnidad?: string | null;
  turno?: string;
  ultimoNewsan?: number | null;
  ultimaFechaEscaneado?: string | null;
  producidoPlacas?: number | null;
  paradasPlacas?: number | null;
  sync?: boolean;
}